+++ packages/react-habitat-redux/src/index.js
@@ -1,0 +1,4 @@
+import ReduxDomFactory from './ReduxDomFactory';
+
+export { ReduxDomFactory };
+export default ReduxDomFactory;