+++ packages/react-habitat-redux/tests/fixtures/MockActions.js
@@ -1,0 +1,7 @@
+
+export const increment = (amount) => ({
+	type: 'INCREMENT',
+	amount,
+});
+
+export default increment;