+++ packages/react-habitat-redux/tests/fixtures/mockReducer.js
@@ -1,0 +1,9 @@
+
+export default function counter(state = { count: 0 }, action) {
+	switch (action.type) {
+	case 'INCREMENT':
+		return { count: state.count + action.amount };
+	default:
+		return state;
+	}
+}